requirements.txt
python-dotenv>=1.0.1
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
requests>=2.32.0
beautifulsoup4>=4.12.3
tqdm>=4.66.5
pillow>=10.4.0
pdf2image>=1.17.0
pytesseract>=0.3.13
pdfminer.six>=20231228
pydantic>=2.8.2
click>=8.1.7
playwright>=1.47.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
numpy>=1.24.0
flask>=2.2.0
spacy>=3.6.0
dateparser>=1.1.8
networkx>=3.0
openai>=1.44.0
markdown>=3.6
faiss-cpu>=1.7.4
